+++ db/migrations/20230330053939_create_demoday_registrations.sql
@@ -1,7 +1,7 @@
 -- migrate:up
 CREATE TABLE demoday_registrations (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    user_id INT NOT NULL,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT NOT NULL,
     demoday_id BIGINT NOT NULL,
     registration_time TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     CONSTRAINT fk_demoday_registrations_user_id FOREIGN KEY (user_id) REFERENCES users(id),
